Add sub-optimal fix for Qt5 wrapping, restoring the previous behavior.
nixpkgs now requires explicitly enabling or disabling Qt5 wrapping. This commit
disables wrapping by default, which breaks Qt apps that aren't overridden
manually in distro-overlay.nix, but fixes all the Qt libraries that don't need
wrapping. This is a sub-optimal solution, but matches the behavior that was
there before wrapping was made mandatory.