Use Python 3.8
A change to boost is needed to build lanelet2-python: components such
as boost python and boost python numpy are used by lanelet2-python,
whose build relies on MRT for dependency discovery. MRT looks for
library files with names like `libboost_python3.so`, so we arrange
that such a name is a symlink to a more specifically-versioned
library, such as `libboost_python38.so` when we are building with
python-3.8.