buildEnv: add Python packages to environment and set LD_LIBRARY_PATH
Now that we set PYTHONPATH in the wrappers, we can add Python packages to the
environment as well. This is necessary to make certain ros2cli commands work
(e.g. ros2launch).

Also, ROS2 loads the RMW implementation dynamically, so we need to set
LD_LIBRARY_PATH in the wrappers.