Configure the Cachix binary cache in flake.nix
With this, Flake users will be automatically asked whether they want
to use the binary cache and its key. If they answer yes, the cache
will be automatically used without any needed configuration.

Additionally, the users will be asked whether they want to mark this
settings as permanent. In that case the cache will be used
automatically in the future without any interaction.

This means that when anybody runs a command like this:

    nix shell github:lopsided98/nix-ros-overlay#humble.rviz2

they will automatically get the packages from the binary cache.