Update ros2-basic example to include ros-core and colcon
Previously, ros-core was not included presumably due to its dependency
on ros2doctor, which depended on ifcfg package unavailable in nixpkgs.
Since ros2cli version 0.17.0 (2022-01-25) ifcfg is no longer necessary
as it was replaced by psutils [1]. This means that we can use ros-core
without problems instead of manually specifying its dependencies such
ros-environment, ros2topic, etc.

In addition to ros-core, colcon is a basic tool for ROS 2. Include it
in the example too.

[1]: https://github.com/ros2/ros2cli/commit/cc5d6b1f9b26ada1a8f366ef7fe9534077c859c2