Allow using buildEnv as ROS underlay
When one has a package X in a Nix development environment (or
nix-shell) and a package of the same name appears in a local ROS
workspace, previous buildEnv implementation ignored the local package.
With this change, the user can set `underlay = true` to make the local
package override the package from Nix. In that case, if the ROS
workspace has no package named X, the one from Nix will be visible.

Originally I thought, it should be possible to achieve the same
without modifying nix-ros-overlay, by just disabling wrapping of
binaries (e.g. ros2) in the buildEnv with `wrapPrograms = false`. But
for some reason, it doesn't work. It makes some packages not visible
at all, which looks like a bug. Until this bug is fixed, we just do
what's in this commit.

Moreover, it's not sufficient to change only AMENT_PREFIX_PATH. It
seems that `ros2 launch` searches the packages based on PYTHONPATH
rather than based on AMENT_PREFIX_PATH. Therefore, changes performed
in Python code in a local workspace launched via `ros2 launch` were
not visible. We fix that by setting also PYTHONPATH.

And while we're at it, we also change the remaining variables. It
seems to be the right thing for common use cases.