rmw-implementation: Don't build with CycloneDDS support on MacOS
CycloneDDS depends on Linux-only libraries such as acl and cannot be
built on MacOS. Removing CycloneDDS from rmw-implementation
dependencies is OK, because there are other RMW implementations, which
build fine on MacOS. For example FastDDS, which is the default one.

The result of this change is that on MacOS one cannot switch to
CycloneDDS at run-time by selecting it in the RMW_IMPLEMENTATION
environment variable.

Specifically, this commit avoids to following error:

       error: Package ‘acl-2.3.1’ in
       /nix/store/.../pkgs/development/libraries/acl/default.nix:40 is
       not available on the requested hostPlatform:
       hostPlatform.config = "aarch64-apple-darwin"