humble.rosidl-generator-py: Fix numpy include path
In the past NumPy core headers were under numpy/core/include, but in
recent nixpkgs versions, the location changed to numpy/_core/include.

NumPy >= 2.0 also includes numpy-config tool. We could use the tool to
get the correct path, instead of hardcoding it in the Nix expression,
but I'm not sure how it would work with cross-compiling, so I'm
leaving the logic as it was.