Update patchAmentVendorGit to use automatically extracted information about vendored source
The information is stored in vendored-source.json files in the package
directories. The file can be automatically generated by running
command like this:

    $(nix-build -A rosPackages.jazzy.gz-math-vendor.updateAmentVendor)'

This command executes Cmake for the given package with an alternative
implementation of the ament_vendor Cmake macro, which generates the
file instead of fetching the source code. The above command generates
`distros/jazzy/gz-math-vendor/vendored-source.json` files with this
content:

    {
      "url": "https://github.com/gazebosim/gz-math.git",
      "rev": "gz-math7_7.5.2",
      "hash": "sha256-LwYeyv8nwX06n5ju+ra2uqNMedMSLRumem8qDHXtNns="
    }

patchAmentVendorGit function is updated to pick the information about
patching from that file instead of from its arguments.

patchGzAmentVendorGit is updated similarly and the check for version
mismatch is removed, because it would not work (version information is
not available from function arguments) and more importantly, it
is (hopefully) no longer needed, because the version will always be
updated automatically by CI.