workflows: Automatically update ament_vendor metadata after running superflore
The update workflow is updated as follows: Instead of running just
superflore, we run:

1. superflore --dry-run ...
2. maintainers/scripts/update-ament-vendor.sh && git commit
3. superflore --pr-only ...

The first command updates the overlay as before, but does not submit a
PR. The second step updates ament_vendor metadata according to the
updated overlay and the third command creates the PR with all updates.

I tried to make the step 2 as fast and efficient as possible, but it
may happen that it will run for long time (about 1 hour in my
testing). The reason is that, for example, if gz-msgs-vendor package
gets updated, the update of other vendored packages depending on it,
e.g., gz-sim-vendor, will have to build the updated gz-msgs-vendor,
because it is not yet in the cachix cache.