Allow generating vendored-source.json for packages with multiple ament_vendor calls
This is needed at least for zenoh-cpp-vendor, which is being prepared
in #558.

The format of vendored-source.json is changed from a single object to
an array of objects. The order of objects corresponds to the order of
ament_vendor calls in the patched CMakeLists.txt.

ament_cmake_vendor_packageConfig.cmake was updated to generate such a
format automatically and patchAmentVendorGit was updated to understand
that format and patch all call sites in the correct order.

The vendored-source.json files are regenerated automatically and there
are no changes except addition of extra brackets.