Allow generating vendored-source.json for packages with multiple ament_vendor calls
This is needed at least for zenoh-cpp-vendor, which is being prepared
in #558, but it also improves the robustness of the whole patching
machanism.

The format of vendored-source.json is changed from a single object to
object of objects, one for each call to ament_vendor. The key in the
top-level object is the target name specified in the ament_vendor call
in CMakeLists.txt.

The principle of operation is slightly different from the previous
version. Instead of introducing "fake" ament_cmake_vendor_package just
for creation of vendored-source.json, we introduce a "wrapped" version
of ament_cmake_vendor_package, which is used both for creation of
vendored-source.json as well as for compiling the package. When
compiling the package, the wrapped version captures ament_vendor calls
and modifies the arguments to use the prefetched data from the Nix
store instead of downloading them directly.

patchAmentVendorGit was updated to understand the new format and to
invoke CMake with the wrapped package and information from
vendored-source.json.

The vendored-source.json files will be changed to the new format in
the next commit.