zenoh-cpp-vendor: Prepare for cross-compilation
Without this change, it is not possible to cross-compile
zenoh-cpp-vendor. However, as of now, cross-compiling doesn't work
even with this commit, because our nixpkgs version is too old and does
not include https://github.com/NixOS/nixpkgs/commit/112d13c5c7c83593972e786cfa9f6fcef06bdfd9c.
So this change is just to be ready for cross-compilation support it after
we move to nixpkgs which includes the mentioned commit.