rolling.foxglove-bridge: Fetch SDK via Nix rather than via cmake FetchContent
Fetching via cmake fails in Nix sandbox so we have to patch it.

For now, we have to keep the SDK version and hash is sync with
CMakeLists.txt. Let's see how often it will change. If too often, we
may think about automating the updates similarly as in
patchAmentVendorGit.