patchExternalProjectGit: Log correct revision in case of mismatch
Previously, if patchExternalProjectGit failed due to revision
mismatch, one had to unpack the original source code and look into
CMakeLists.txt file which revision is used there. With this change,
the correct revision will be printed as a part of the error message in
most cases.

An example of printed error message is:

    patchExternalProjectGit: did not find revision: b63a0b6f79d3ea91dc221724b42dae49894449fc
    (found fb735b79cab5f0cdda45bc5087414d405ef8f3ab)

This updates the case when revVariable == "". The other case will be
updated in the next commit.