colcon: Downgrade setuptools dependency to version 79
colcon_core uses functionality of setuptools to implement
`colcon build --symlink-install` for Python packages. Setuptools
version 80 removes parts of this functionality [1][2]. More
specifically, colcon invokes setup.py of the package to be installed
as follows:

   setup.py develop --editable --build-directory build --no-deps

and with setuptools >= 80, this fails with `error: option --editable
not recognized`. For this reason, colcon_core now requires
setuptools < 80 [3].

AFAICT there is no easy way to replace this setuptools functionality
with something else now. Eventually, colcon-python-project [3] could
be the solution, but until it's mature, downgrading setuptools seems
to be the way to go.

Invocation of setup.py happens in colcon-core [4] so we have to
downgrade setuptools in this package. With it, colcon-python-setup-py
complains about two different setuptools versions, so we downgrade
there as well.

[1]: https://setuptools.pypa.io/en/stable/history.html#v80-0-0
[2]: https://github.com/colcon/colcon-core/issues/696
[3]: https://github.com/colcon/colcon-python-project
[4]: https://github.com/colcon/colcon-core/blob/3310754d1eeb820368580d569901ba4d26166717/colcon_core/task/python/build.py#L154