lib: Add patchAmentVendorFile
This is almost the same as patchAmentVendorGit but uses fetchurl
instead of fetchgit to download the required data.

Due to the similarity with patchAmentVendorGit, the common
functionality was moved to patchAmentVendor and patchAmentVendorGit
and patchAmentVendorFile are thin wrappers around it.

This function will be used in a subsequent commit to patch
clips-vendor.