vtk, pcl, rtabmap: Switch to Qt6
In the past, vtk was overridden to use Qt5 to support rtabmap
build (#639). However, by using qt5.full, it adds dependency on
insecure qtwebengine to packages like pcl, which is used by many ROS
packages including ros-desktop metapackages, which now fail to
evaluate. Additionally, Qt5 is no longer supported [1], so we need to
start switching everything to Qt6.

This is just a first step to get rid of most important evaluation
failures. Many things now evaluate but still fail to build. This will
need to be addressed later.

[1]: https://www.qt.io/blog/extended-security-maintenance-for-qt-5.15-begins-may-2025