Build ROS packages with Qt6 instead of Qt5
Some packages fail to build with Qt5 and the remaining packages then
complain with "Error: detected mismatched Qt dependencies" when some
of their dependencies are built with Qt5 and other with Qt6. Let's
switch everything to Qt6.

Note that while newer ROS distros (currently Rolling) introduce new
qt6 rosdep keys, older distros are unlikely to be migrated to Qt6 so
we have to pretend that qt6 is called qt5.