Update MRPT hash
This time, humble stayed at the old version due to "held for sync" [1]
so we need to include two hashes. Alternative solution would be to
copy to distro-specific overlays, but this would be too much work
later when all the distros are again in sync. Also, the planned MRPT 3
release will probably make this override unnecessary.

[1]: https://github.com/ros/rosdistro/pull/48722#pullrequestreview-3492252931