ogre1_9: Build without FreeImage
FreeImage is insecure and was removed from nixpkgs (but we still use
an older nixpkgs version with freeimage). To allow building gazebo
without insecure packages, let's build ogre without it. No package
fails to build, but this might cause problems at runtime. From my
simple experiments with gazebo, it seems it's not the case.