nix-ros-build-action: Don't push sources to the binary cache
It seems that our binary cache is full and we're building a lot of
derivations every CI run again and again. The reason might be (besides
the low cache size limit) that we're storing in the cache both
derivation sources and outputs. In an aim to make CI and binary cache
more reliable, this commit changes that and from now on, we will cache
only the outputs.

This means that when some source disappears or gets renamed (both
happened in the past) our users will not be able to rebuild all
packages from source. Hopefully, this will not happen too often. And
even if it happens, it's probably less problematic then when users
have to build a lot of packages themselves when they just want to use
ROS.