nix-ros-build-action: Allow pinning the result in the Cachix cache
This should ensure that the master and develop branches are not
evicted from the cache due to the pushes to other branches.

In addition to that, the action now prints size statistics to the log.
Hopefully, it will allow understanding how is the cache used and
whether we can afford to put everything there given its limited
size (in the free plan).