superflore: Generate Nix derivations even with unresolved dependencies
Previously, packages with unresolved dependencies were not generated
at all. This was quite confusing for users, because it was not clear
why a package is missing. It was necessary to look at the last
superflore PR as well as the log from the GitHub Action run to
understand which dependency is missing.

With this change, unresolved dependencies are used in the generated
derivations as their rosdep keys prefixed by "_unresolved_". The end
effect should be that when users try to build their favorite package,
they'll see errors like:

   lib.customisation.callPackageWith: Function called without required
   argument "_unresolved_FOO" at /nix-ros-overlay/distros/rolling/BAR/default.nix:5

This error is more actionable than "attribute 'BAR' not found".